StrahlkorperCoordsInDifferentFrame: time-indep map
StrahlkorperCoordsInDifferentFrame now handles the case where
a block is time-independent, but the input and output coords
have different Frame Tags, e.g. one is Inertial and the other
is Grid.  This can happen even though (for the time-independent
case) Grid, Inertial, and Distorted are really the same frame.

Previously StrahlkorperCoordsInDifferentFrame threw a
runtime error in that case.