Check dense output ability before postprocessors
If the TimeStepper knows it cannot produce dense output without
another substep, there is no reason to wait for the postprocessors.
This is particularly important if the step size can be adjusted later
in the action loop, as that could push triggers into the next step.

Changing the step size still does not work properly with
AdamsBashforth, as it is always able to produce dense output.